sqlite/template/requirements/development.txt
Django==2.2.10
pytz==2019.3
sqlparse==0.3.0
pre-commit==2.0.1
black==19.10b0
flake8==3.7.9
freezegun==0.3.14
django-extensions==2.2.8
ipython==7.12.0
ipython-genutils==0.2.0
django-split-settings==1.0.1
